1. Отладка записи конфигурации, так как перетералась стаой версией package.json

+++ src/components/mono_repository/build.ts
@@ -9,39 +9,41 @@ export async function build(CWD: string) {
   console.log(chalk.bold.green(`[ ${pkg.name} ][ BUILD ]`));
   console.log("");
 
-  const result = spawn.sync("npm", ["run", "build"], {
-    cwd: CWD,
-    encoding: "utf8",
-    stdio: "inherit",
-  });
+  if (envConfig.isDevelopment) {
+    const result = spawn.sync("npm", ["run", "build"], {
+      cwd: CWD,
+      encoding: "utf8",
+      stdio: "inherit",
+    });
 
-  if (result.signal) {
-    if (result.signal === "SIGKILL") {
-      console.log(
-        chalk.red(
-          "The build failed because the process exited too early. " +
-            "This probably means the system ran out of memory or someone called `kill -9` on the process.",
-        ),
-      );
+    if (result.signal) {
+      if (result.signal === "SIGKILL") {
+        console.log(
+          chalk.red(
+            "The build failed because the process exited too early. " +
+              "This probably means the system ran out of memory or someone called `kill -9` on the process.",
+          ),
+        );
 
-      process.exit(1);
-    } else if (result.signal === "SIGTERM") {
-      console.log(
-        chalk.bold.red(
-          "The build failed because the process exited too early. " +
-            "Someone might have called `kill` or `killall`, or the system could be shutting down.",
-        ),
-      );
+        process.exit(1);
+      } else if (result.signal === "SIGTERM") {
+        console.log(
+          chalk.bold.red(
+            "The build failed because the process exited too early. " +
+              "Someone might have called `kill` or `killall`, or the system could be shutting down.",
+          ),
+        );
 
-      process.exit(1);
+        process.exit(1);
+      }
+
+      process.exit(0);
     }
-
-    process.exit(0);
   }
 
   if (!envConfig.isDevelopment) {
     // tslint:disable-next-line:variable-name
-    const result_release = spawn.sync("npm", ["run", "build:release"], {
+    const result_release = spawn.sync("npm", ["run", "build:both"], {
       cwd: CWD,
       encoding: "utf8",
       stdio: "inherit",
